requirements.txt
astroid==3.1.0
cfgv==3.4.0
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
filelock==3.13.1
greenlet==3.0.3
identify==2.5.35
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
psycopg2==2.9.9
pylint==3.1.0
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
setuptools==69.2.0
SQLAlchemy==2.0.29
tomlkit==0.12.4
typing_extensions==4.10.0
virtualenv==20.25.1
